requirements.txt
annotated-types==0.5.0
anyio==3.7.1
Authlib==1.2.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
dnspython==2.4.1
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
numpy==1.24.4
pandas==2.0.3
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
attrs==23.1.0
boto3==1.28.14
botocore==1.31.14
cachetools==5.3.1
charset-normalizer==3.2.0
ecdsa==0.18.0
email-validator==2.0.0.post2
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
oauthlib==3.2.2
orjson==3.9.2
outcome==1.2.0
packaging==23.1
Pillow==10.0.1
pluggy==1.2.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pyhtml2pdf==0.0.7
pyparsing==3.1.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==7.4.0
pytest-mock==3.11.1
python-jose==3.3.0
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
selenium==4.14.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
webdriver-manager==4.0.1
wsproto==1.2.0